feat: add form doc and example
