+++ app/src/components/Home.vue
@@ -42,6 +42,21 @@
           </tr>
 
           <tr>
+            <td>form</td>
+            <td>
+              <router-link to="/form" target="_blank">/form</router-link>
+            </td>
+            <td>
+              <a
+                href="https://github.com/dvisionlab/ditto/blob/master/app/library/form/README.md"
+                target="_blank"
+              >
+                docs
+              </a>
+            </td>
+          </tr>
+
+          <tr>
             <td>http</td>
             <td><router-link to="/http" target="_blank">/http</router-link></td>
             <td>
